packages/etl/requirements.txt
supabase==2.4.0
anthropic==0.25.0
requests==2.31.0
geopandas==0.14.3
shapely==2.0.3
overpy==0.7.0
googlemaps==4.10.0
pandas==2.2.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
tqdm==4.66.2
